+++ CHANGELOG.md
@@ -4,6 +4,8 @@ BUG FIXES:
 * ydb: fixed panic if database was removed outside of terraform
 * storage: fix issue when error, returned from reading extend bucket settings treated as important.
   It will be printed as log output now instead of interrupting plan execution.
+* greenplum: fix bug with not setting `security groups_ids` in `resource_yandex_mdb_greenplum_cluster`
+* greenplum: fix bug with not setting `access` in `resource_yandex_mdb_greenplum_cluster`
 
 ENHANCEMENTS:
 * mdb: add `sqlcollation` attribute to `yandex_mdb_sqlserver_cluster` resource and data source
@@ -12,6 +14,7 @@ ENHANCEMENTS:
 FEATURES:
 * k8s: add `instance_template.name` attribute to `node group` resource and data source
 * k8s: add `instance_template.ipv4_dns_records`, `instance_template.ipv6_dns_records` attributes to `node group` resource and data source
+* greenplum: add fields `greenplum_config` and `pooler_config` in `resource_yandex_mdb_greenplum_cluster`
 
 ## 0.76.0 (July 01, 2022)
 BUG FIXES: